Release Candidate 0.3.0
# Proxy Changelog for Release Candidate 0.3.0
## Update Informations
In this Release Candidate we add a new container with improved configurations, updated the old one for a bugfix and improved the push and tagging scripts.
## General Changes
We changed the file name convention for the NGINX configuration setup. In 1.0.0 and 1.0.1 we had two NGINX configuration files one for HTTPS and one for HTTP only. Now we have only one HTTPS file.
At build time a self-signed cert is now created. Therefore a HTTP only configuration file is not more required.
Please make sure to disable the NGINX configuration volume to get the new file. Or make a backup. Delete all and start from scratch and restore the backup.


## Fixes & Improvements
- Fix Healthcheck of 1.0.0 and 1.0.1
- Improve push.sh, tagging.sh and notify_hub.docker.com.sh to support internal gitlab-ci
- Fix IP restriction file on 1.0.1
- Add new container 1.0.2

## Detailed Changes
- Fix Healthcheck of 1.0.0 and 1.0.1
  curl Package was not found. Now it is installed at build time.
- Improve push.sh, tagging.sh and notify_hub.docker.com.sh to support internal gitlab-ci
  We change the shell scripts to support a second docker registry. Now the scripts look for a variable with the name `INTERNAL_REGISTRY_HOST`.
- Fix IP restriction file on 1.0.1
  `allow 0.0.0.0;` led to a mistake. The new value is: `allow all;`
- Add new container 1.0.2
  - We add a new container, for new LABEL structure and additional VERSION and NAME environment variable.
  - We prepare our proxy for a monitoring solution in the future
  - We renamed the proxy_vars file
  - We change the logfile path to stdout and stderr